Refactor: member domain에 있는 coupon 관련 클래스들 패키지 이동
